requirements.txt
discord.py
aiohttp
requests
beautifulsoup4
colorama
flask
pystyle
faker
python-dotenv